+++ cmake/modules/gluecodium/gluecodium/GetTargetCompileDefinitions.cmake
@@ -68,9 +68,11 @@ function(gluecodium_get_target_compile_definitions _target)
 
     if("COMMON" IN_LIST _source_sets)
       set(_sync_cache_property "$<TARGET_PROPERTY:${_target},GLUECODIUM_SYNCHRONIZE_ACCESS_CLASS_CACHE>")
+      set(_enable_internal_debug_check "$<TARGET_PROPERTY:${_target},GLUECODIUM_ENABLE_INTERNAL_DEBUG_CHECKS>")
       list(APPEND _public $<$<NOT:$<STREQUAL:${_common},${_main}>>:${_common}_SHARED>)
       list(APPEND _private $<$<NOT:$<STREQUAL:${_common},${_main}>>:${_common}_INTERNAL>
-                           $<$<BOOL:${_sync_cache_property}>:GLUECODIUM_SYNCHRONIZE_ACCESS_CLASS_CACHE>)
+                           $<$<BOOL:${_sync_cache_property}>:GLUECODIUM_SYNCHRONIZE_ACCESS_CLASS_CACHE>
+                           $<$<BOOL:${_enable_internal_debug_check}>:GLUECODIUM_ENABLE_INTERNAL_DEBUG_CHECKS>)
     endif()
 
     if(_args_RESULT_PUBLIC)